requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
lxml==4.9.1
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
pytz==2022.7
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.17.1
